+++ CHANGELOG.rst
@@ -1,6 +1,22 @@
 Toolium Changelog
 =================
 
+v1.1.0
+------
+
+*In development*
+
+- New MobilePageObject class to test Android and iOS apps with the same base page object
+- Fix visual report links in Windows
+- Add @no_reset_app, @reset_app and @full_reset_app behave tags to configure Appium reset capabilities for one scenario
+- Add @android_only and @ios_only behave tags to exclude one scenario from iOS or Android executions
+- Add a behave user property named *env* to select config files, e.g., to use android-properties.cfg file:
+
+.. code:: console
+
+    $ behave -D env=android
+
+
 v1.0.1
 ------